Duplicate the admin_analytics_tests and fix the ones using the GOV.UK Design System
